+++ src/main/kotlin/com/atlassian/performance/tools/virtualusers/api/VirtualUserLoad.kt
@@ -10,7 +10,7 @@ import java.time.Duration
  * [ramp] duration when load is ramping up
  * [flat] duration when full load is generated
  */
-data class VirtualUserLoad(
+class VirtualUserLoad(
     val virtualUsers: Int = 10,
     val hold: Duration = Duration.ZERO,
     val ramp: Duration = Duration.ofSeconds(15),